requirements.txt
numpy==1.14.2
scipy==1.0.1
pillow==5.1.0
toolz==0.9.0
pyrsistent==0.14.2
torchvision==0.2.0
tensorboardX
tqdm==4.23.1